catch tx error
